♻️ JE-105-refactor: generateCalendar함수 컴포넌트화
