+++ src/components/mypage/BookingCalendar/BookingCalendar.tsx
@@ -1,14 +1,15 @@
 import useClickOutside from '@/hooks/useClickOutside';
 import {
-  getDaysInMonth,
-  getFirstDayOfMonth,
   getLocalDateString,
+  isPastDate,
+  removeTime,
 } from '@/libs/utils/dateUtils';
 import { GetMyActivitiesReservationDashboardResponse } from '@trip.zip-api';
 import { useRef, useState } from 'react';
 
 import BookingDetailModal from './BookingDetailModal';
 import StatusTag from './BookingStatusTag';
+import { generateCalendar } from './generateCalendar';
 
 type CalendarProps = {
   currentYear: number;
@@ -17,26 +18,6 @@ type CalendarProps = {
   monthlyData: GetMyActivitiesReservationDashboardResponse;
   activityId: number;
   onRefresh: () => void;
-};
-
-type DateObject = {
-  date: Date;
-  day: number;
-  isCurrentMonth: boolean;
-  bookingInfo?: {
-    completed: number;
-    confirmed: number;
-    pending: number;
-  };
-};
-
-const removeTime = (date: Date) => {
-  date.setHours(0, 0, 0, 0);
-  return date;
-};
-
-const isPastDate = (date: Date, today: Date) => {
-  return removeTime(new Date(date)) < today;
 };
 
 export default function Calendar({
@@ -52,97 +33,25 @@ export default function Calendar({
 
   const today = removeTime(new Date());
 
-  // 현재 달의 날짜 계산
-  const firstDayOfCurrentMonth = getFirstDayOfMonth(currentYear, currentMonth);
-  const daysInCurrentMonth = getDaysInMonth(currentYear, currentMonth);
-
-  // 이전 달의 날짜 계산
-  const prevMonth = currentMonth === 0 ? 11 : currentMonth - 1;
-  const prevYear = currentMonth === 0 ? currentYear - 1 : currentYear;
-  const daysInPrevMonth = getDaysInMonth(prevYear, prevMonth);
-
-  // 다음 달의 날짜 계산
-  const nextMonth = currentMonth === 11 ? 0 : currentMonth + 1;
-  const nextYear = currentMonth === 11 ? currentYear + 1 : currentYear;
-
   // 날짜 별 예약정보 리스트업
-  const bookingMap = new Map<
-    string,
-    { completed: number; confirmed: number; pending: number }
-  >();
+  const bookingMap = new Map<string, Record<string, string>>();
+
   monthlyData.forEach((booking) => {
-    bookingMap.set(booking.date, booking.reservations);
+    const transformedReservations = {
+      completed: String(booking.reservations.completed),
+      confirmed: String(booking.reservations.confirmed),
+      pending: String(booking.reservations.pending),
+    };
+
+    bookingMap.set(booking.date, transformedReservations);
   });
 
-  // 캘린더 생성 함수
-  const generateCalendar = () => {
-    const calendar: DateObject[][] = [];
-    let week: DateObject[] = [];
-
-    // 이전 달 날짜 채우기
-    for (let i = firstDayOfCurrentMonth - 1; i >= 0; i--) {
-      const date = new Date(prevYear, prevMonth, daysInPrevMonth - i);
-      const dateString = getLocalDateString(date);
-      week.push({
-        date,
-        day: daysInPrevMonth - i,
-        isCurrentMonth: false,
-        bookingInfo: bookingMap.get(dateString),
-      });
-    }
-
-    // 현재 월 날짜 채우기
-    for (let day = 1; day <= daysInCurrentMonth; day++) {
-      const date = new Date(currentYear, currentMonth, day);
-      const dateString = getLocalDateString(date);
-      week.push({
-        date,
-        day,
-        isCurrentMonth: true,
-        bookingInfo: bookingMap.get(dateString),
-      });
-      if (week.length === 7) {
-        calendar.push(week);
-        week = [];
-      }
-    }
-
-    // 다음 달 날짜 채우기
-    let nextMonthDay = 1;
-    while (week.length < 7) {
-      const date = new Date(nextYear, nextMonth, nextMonthDay);
-      const dateString = getLocalDateString(date);
-      week.push({
-        date,
-        day: nextMonthDay,
-        isCurrentMonth: false,
-        bookingInfo: bookingMap.get(dateString),
-      });
-      nextMonthDay++;
-    }
-    calendar.push(week);
-
-    // 나머지 빈 줄 채우기
-    while (calendar.length < 6) {
-      week = [];
-      for (let i = 1; i <= 7; i++) {
-        const date = new Date(nextYear, nextMonth, nextMonthDay);
-        const dateString = getLocalDateString(date);
-        week.push({
-          date,
-          day: nextMonthDay,
-          isCurrentMonth: false,
-          bookingInfo: bookingMap.get(dateString),
-        });
-        nextMonthDay++;
-      }
-      calendar.push(week);
-    }
-
-    return calendar;
-  };
-
-  const calendar = generateCalendar();
+  // 캘린더 생성
+  const calendar = generateCalendar({
+    currentYear,
+    currentMonth,
+    dataMap: bookingMap,
+  });
 
   const handleDateClick = (date: string) => {
     // 예약 정보가 있을 때만 모달 열기
@@ -179,7 +88,7 @@ export default function Calendar({
             ? 'bg-custom-gray-800'
             : 'bg-green-400';
           const dateString = getLocalDateString(dateObject.date);
-          const hasBooking = !!dateObject.bookingInfo;
+          const hasBooking = !!dateObject.scheduleInfo;
           return (
             <button
               key={`${weekIndex}-${dateIndex}`}
@@ -196,12 +105,12 @@ export default function Calendar({
                 <span className="text-17 font-medium md:text-21">
                   {dateObject.day}
                 </span>
-                {dateObject.bookingInfo && (
+                {dateObject.scheduleInfo && (
                   <div className={`${alertClass} size-8 rounded-full`} />
                 )}
               </div>
-              {dateObject.bookingInfo && (
-                <StatusTag bookingInfo={dateObject.bookingInfo} />
+              {dateObject.scheduleInfo && (
+                <StatusTag bookingInfo={dateObject.scheduleInfo} />
               )}
             </button>
           );
